Fix _WeightedLeastSquaresPWLSolver for underdetermined systems.
The solver uses np.linalg.lstsq for non-mono linear least squares. Normally, .lstsq returns residuals from which we compute the squared error, but in the underdetermined case it omits the residuals. Our solver previously assumed a squared error of 0 in the underdetermined case. In this CL, we construct a test of an underdetermined case where the squared error is not 0, and modify the solver to manually calculate the squared error in this case.

PiperOrigin-RevId: 321287585
Change-Id: Ia617500a1626fc5a3a1f528d051cbe8a980cee91