Convert pwlcurve points and inputs to float64 to avoid precision-based boundary glitches.
PWLCurve evaluation assumes that transform(min_knot_x) <= transform(clipped(x)) <= transform(max_knot_x) for all x in the data, since we assume that transform is a monotonically increasing function. Unfortunately, the inequality can fail if x uses a different floating point precision than the knot-xs.

This CL tests that case with an example that previously failed, and casts the types appropriately so that it no longer fails.

PiperOrigin-RevId: 324734893
Change-Id: I9f2be5454aa18555035eea67224f78a37df098ac