Add an option to /pwlfit/ that disables learning endpoints.
Curves fit better if we learn endpoints. However, learned endpoints cause flat sections to the left and right of the curve, even if the user set min_slope > 0 or max_slope < 0. In some cases it's desirable for min/max slope to extend all the way to the extremes. To support those cases, we add a 'learn_ends' boolean parameter that can be disabled to avoid learning endpoints.

The motivating use case is 'strict monotonicity', which can now be obtained by setting min_slope = epsilon (or max_slope = -epsilon) and learn_ends = False.

PiperOrigin-RevId: 345162650
Change-Id: I4e8a48945df60e5494fe4d4930b490c2dade82d1