Add bitonic curve fitting to /pwlfit/ as a compromise between mono and nonmono fitting.
A nonmono curve is unrestricted -- its slope can change direction at any of its control points. A monotone curve is tightly restricted -- its slope can never change direction. A bitonic curve allows precisely one direction change, which can be a useful compromise for some kinds of data. A bitonic curve will have a peak (if the slope is initially positive and then switches to negative) or a trough (if the slope is initially negative and then switches to positive). Bitonic curves are a good fit if that pattern is expected and desired. (A nonmono fitting might learn a bitonic curve, but it might also learn additional zigs and zags.)

Bitonic curves are well-suited for features like body temperature or blood sugar, where extreme lows and extreme highs are dangerous and the ideals are usually in the middle. A monotone curve would be unable to fit the underlying pattern, whereas an unrestricted nonmono curve might overfit to noisy parts of the data.

PiperOrigin-RevId: 347859594
Change-Id: Ib2a139e131f47b58fa1417b4fcace729fe859f72